bot/requirements.txt
discord.py==2.3.2
httpx==0.26.0
python-dateutil==2.8.2
